Use ModuleClass to load views
