+++ src/machinable/storage/experiment.py
@@ -9,7 +9,7 @@ from .collections import ComponentStorageCollection, ExperimentStorageCollection
 from .component import StorageComponent
 from .models import StorageExperimentModel
 from .models.filesystem import StorageExperimentFileSystemModel
-from .view import _bind_view
+from .view import View
 
 
 class StorageExperiment:
@@ -198,7 +198,7 @@ class StorageExperiment:
     @property
     def view(self):
         """Returns the registered view"""
-        return _bind_view("experiment", self)
+        return View.bind("experiment", self)
 
     def __len__(self):
         """Returns the number of components in this experiment"""
